+++ packages/search-in-workspace/src/browser/search-in-workspace-result-tree-widget.tsx
@@ -42,6 +42,7 @@ import { SearchInWorkspaceService } from './search-in-workspace-service';
 import { MEMORY_TEXT } from './in-memory-text-resource';
 import URI from '@theia/core/lib/common/uri';
 import * as React from 'react';
+import { SearchInWorkspacePreferences } from './search-in-workspace-preferences';
 
 const ROOT_ID = 'ResultTree';
 
@@ -115,6 +116,7 @@ export class SearchInWorkspaceResultTreeWidget extends TreeWidget {
     @inject(LabelProvider) protected readonly labelProvider: LabelProvider;
     @inject(WorkspaceService) protected readonly workspaceService: WorkspaceService;
     @inject(TreeExpansionService) protected readonly expansionService: TreeExpansionService;
+    @inject(SearchInWorkspacePreferences) protected readonly searchInWorkspacePreferences: SearchInWorkspacePreferences;
 
     constructor(
         @inject(TreeProps) readonly props: TreeProps,
@@ -152,6 +154,10 @@ export class SearchInWorkspaceResultTreeWidget extends TreeWidget {
 
         this.toDispose.push(this.editorManager.onActiveEditorChanged(() => {
             this.updateCurrentEditorDecorations();
+        }));
+
+        this.toDispose.push(this.searchInWorkspacePreferences.onPreferenceChanged(() => {
+            this.update();
         }));
     }
 
@@ -588,6 +594,7 @@ export class SearchInWorkspaceResultTreeWidget extends TreeWidget {
     protected renderResultLineNode(node: SearchInWorkspaceResultLineNode): React.ReactNode {
         const prefix = node.character > 26 ? '... ' : '';
         return <div className={`resultLine noWrapInfo ${node.selected ? 'selected' : ''}`}>
+            {this.searchInWorkspacePreferences['search.lineNumbers'] && <span className='theia-siw-lineNumber'>{node.line}</span>}
             <span>
                 {prefix + node.lineText.substr(0, node.character - 1).substr(-25)}
             </span>